Solve: [2week] LeetCode_643_Maximum Average Subarray I - 정문영

+++ 2week/LeetCode_643_Maximum Average Subarray I/정문영/Solution.java
@@ -1,8 +1,8 @@
 public class Solution {
 	public double findMaxAverage(int[] nums, int k) {
 		int sum = 0;
-        int cnt = 0;
-        double max_sum = Integer.MIN_VALUE;
+    int cnt = 0;
+    int max_sum = Integer.MIN_VALUE;
 		
 		
 		for (int i=0; i<nums.length; i++) {
@@ -16,6 +16,6 @@ public class Solution {
 			}
 		}
 		
-		return max_sum / k; 
+		return (double)max_sum / k; 
   }
 }
